+++ include/utils/utilities.h
@@ -2,5 +2,6 @@
 #define UTILITIES_H
 
 bool rgu_isStrInArray(const char **arr, int arrSize, const char* str);
+void rgu_trimStr(char *str);
 
 #endif // UTILITIES_H